requirements/base.txt
Django==1.11.26
django-environ==0.4.5
django-tenant-schemas==1.9.0
psycopg2==2.7.6.1
djangorestframework==3.9.4
Pillow==6.1
django-reversion==3.0.4
drf-flex-fields==0.3.4
django-waffle==0.17
django-import-export==1.2.0
dateutils==0.6.6
graphene-django-extras==0.4.6
graphene-django==2.5.0
Pygments==2.3.1
django-sequences==2.3
num2words==0.5.10
graphene-file-upload==1.2.2
pyexcelerate==0.8.0